+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The Decision tree now has a minimum height and weight to make sure it is always shown.
 - `react-craft-ai-operations-history` has now a default estimated period between samples of 1.
 - Position and zoom can now be correctly set in the DecisionTree props.
+- Nodes button are not moving when they are selected.
 
 ## [0.0.24](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.23...v0.0.24) - 2019-07-10