+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.24...HEAD)
+## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.25...HEAD)
+
+## [0.0.25](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.24...v0.0.25) - 2019-07-24
 
 ## Fixed